enable backups

+++ database/migrations/2016_08_18_152401_create_likes_table.php
@@ -18,7 +18,7 @@ class CreateLikesTable extends Migration
             $table->string('likeable_id');
             $table->string('likeable_type');
             // $table->string('item_id');
-            $table->smallInteger('vote');
+            $table->tinyInteger('vote');
             $table->timestamps();
         });
     }
